refactor: divide logic into http and grpc modules
